requirements.txt
aniso8601==9.0.1
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
eventlet==0.33.0
Flask==2.0.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.19.4
pycparser==2.21
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
PyYAML==3.10
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
